+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Fixed Category-Rules performing just a StartsWith and not a Contains search. 
+* Fixed sample data causing extra categories to appear
+* Fixed reports not updating when settings change
+* Fixed bug with categories with different capitalization causing issues